Bug 1548687 - Send OOP frame transforms to the main thread more often. r?botond
In cases such as APZ scrolling with paint-skipping, there is an
additional transform component that needs to be included in the
transforms that APZ sends to the main thread for the purposes of
tracking OOP iframe positions. This component also updates more
frequently, generally on each call to RequestContentRepaint, as
mExpectedGeckoMetrics is updated.

This patch adds some machinery to send partial updates to the main
thread, for layers subtrees that are descendants of a given APZC. It
also includes the missing transform component in the calculated
matrices.