Bug 1561570 - Ensure the BrowserChild is updated with the new chrome offset after fullscreen changes happen. r?rhunt
Apparently the GetChromeOffset() in BrowserChild becomes stale after exiting
fullscreen, but it's not clear to me what is supposed to keep it updated. So
this patch adds a fullscreenchange event listener that updates it. If we
encounter other scenarios where this happens maybe we can find a better way
to do this but without more knowledge of the scenarios this is the best option
I have.