Bug 1617179 - Extract a helper method across WR and non-WR codepaths. r?botond
The operations of recording checkerboard events and advancing the animations
across all APZs are conceptually linked. This patch extracts a helper that
does these operations in a single function, that is then called from both
the WR and non-WR codepaths. This allows us to adjust the locks acquired when
running this code without having to duplicate it in multiple places (see next
patch).