Bug 1617179 - Fold IsCurrentlyCheckerboarding into GetCheckerboardMagnitude. r?botond
The GetCheckerboardMagnitude function is incorrect, and suffers from the same
problem that IsCurrentlyCheckerboarding used to have prior to bug 1617427.
That is, it does not take into account clipping due to the composition bounds
and async transforms of ancestor scrollframes. This patch merges the code
from IsCurrentlyCheckerboarding and GetCheckerboardMagnitude so that there
is a single function (the new GetCheckerboardMagnitude) that takes into account
the ancestor clipping stuff and also returns the checkerboarded area instead
of just a boolean. This allows us to migrate the existing call site of
IsCurrentlyCheckerboarding to use the new function, when then lets us remove
the IsCurrentlyCheckerboarding function entirely.

Since the updated GetCheckerboardMagnitude is called per-composite, this patch
also introduces an additional optimization in order to mitigate any performance
impact. The function now first does the local APZC checkerboard calculation,
which should detect no checkerboarding most of the time. Only in cases where
that does detect checkerboarding, do we fall back to doing the more expensive
recursive checkerboard calculation.